feat(dashboard): add common cell rendering utilities and column helpers
Introduced new utility functions for rendering cells in the dashboard, including badge classes, progress indicators, and signed value formatting. Added column type definitions and helper functions for managing column items. This enhancement streamlines the rendering process and improves the overall structure of the dashboard components.
