+++ apps/app/src/features/dashboard/dashlets/common/column-helpers.ts
@@ -1,0 +1,13 @@
+import type { TableColumn } from "./column-types";
+
+export interface ColumnItem extends TableColumn {
+  _id: string;
+}
+
+export function toColumnItems(columns: TableColumn[]): ColumnItem[] {
+  return columns.map((col, i) => ({ ...col, _id: `col-${i}-${col.key}` }));
+}
+
+export function fromColumnItems(items: ColumnItem[]): TableColumn[] {
+  return items.map(({ key, label, type }) => ({ key, label, type }));
+}
