+++ apps/app/src/features/dashboard/dashlets/common/column-types.ts
@@ -1,0 +1,13 @@
+export type ColumnType = "text" | "badge" | "highlight" | "signed" | "progress";
+
+export interface TableColumn {
+  key: string;
+  label: string;
+  type: ColumnType;
+}
+
+export interface SortConfig {
+  enabled: boolean;
+  /** Ordered list of column keys available in the sort toolbar */
+  columns: string[];
+}
