+++ apps/app/src/features/dashboard/dashlets/data_table/dashlet.settings.tsx
@@ -27,6 +27,8 @@ import {
   defaultSort,
 } from "./dashlet";
 import { SettingsTextField, SettingsSelectField } from "../common";
+import type { ColumnItem } from "../common/column-helpers";
+import { toColumnItems, fromColumnItems } from "../common/column-helpers";
 import AbsoluteModal from "@/features/common/components/absolute-modal/absolute-modal";
 import { tr } from "@/features/i18n/tr.service";
 
@@ -35,22 +37,6 @@ import { tr } from "@/features/i18n/tr.service";
 // ============================================================================
 
 type SettingsTab = "visualization" | "data";
-
-interface ColumnItem extends TableColumn {
-  _id: string;
-}
-
-// ============================================================================
-// Helpers
-// ============================================================================
-
-function toColumnItems(columns: TableColumn[]): ColumnItem[] {
-  return columns.map((col, i) => ({ ...col, _id: `col-${i}-${col.key}` }));
-}
-
-function fromColumnItems(items: ColumnItem[]): TableColumn[] {
-  return items.map(({ key, label, type }) => ({ key, label, type }));
-}
 
 // ============================================================================
 // Component
